core/processing_1_2_0/main.py
"""
Data Processing Module

This module handles cleaning, enrichment, and processing of data from various sources:
- Email cleaning and metadata extraction
- Document text extraction
- Attachment processing
"""

from typing import Dict, Any, List
from bs4 import BeautifulSoup
import docx
from PyPDF2 import PdfReader

class DataProcessor:
    """Handles data processing and cleaning operations."""
    
    def __init__(self):
        """Initialize the data processor with default configurations."""
        # TODO: Load processing configurations
        pass
        
    def process_email(self, raw_email: Dict[str, Any]) -> Dict[str, Any]:
        """Process and clean an email message.
        
        Args:
            raw_email: Raw email data from Graph API
            
        Returns:
            Cleaned and enriched email data
        """
        # TODO: Implement email processing
        pass
        
    def extract_document_text(self, document: Dict[str, Any]) -> str:
        """Extract text content from a document.
        
        Args:
            document: Document data with content
            
        Returns:
            Extracted text content
        """
        # TODO: Implement document text extraction
        pass
        
    def process_attachment(self, attachment: Dict[str, Any]) -> Dict[str, Any]:
        """Process an email attachment.
        
        Args:
            attachment: Attachment data with content
            
        Returns:
            Processed attachment with extracted text and metadata
        """
        # TODO: Implement attachment processing
        pass
