requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.1.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
python-jose==3.3.0
msal==1.24.0
python-multipart==0.0.6
python-docx==0.8.11
pandas==2.1.3
pypdf==4.2.0
python-pptx==0.6.21
openpyxl==3.1.2
chardet==5.2.0
gunicorn==21.2.0
aiofiles==24.1.0
openai==1.78.1
numpy==1.26.4
beautifulsoup4==4.12.3
scikit-learn
pytest-asyncio==0.23.5 